Vtol Parameter Optimisation

Functions

Main.VtolParameterOptimisation.clipping_loss — Method

clipping_loss(x::Real, min::Real, max::Real; offset::Real=100.0)

Returns a quadratic loss at the distance from the minimum or maximum value when the clipping limits are exceeded. An additional offset is added to clearly separate the loss from the rest of the optimisation.

Main.VtolParameterOptimisation.quadratic_prior_loss — Method

quadratic_prior_loss(x::Real, μ::Real, σ::Real)

Returns the squared loss of the value x from the prior μ with the scaling σ. Can be used in the optimisation of an unknown loss landscape to always guarantee a gradient.